replace worker based downloader
